restructure for release as python package add setuptools build
