+++ example/verify_single_card.py
@@ -1,0 +1,14 @@
+from pathlib import Path
+
+from vcv import verify_card
+
+if __name__ == "__main__":
+
+    img_path = str(Path("./example/data/IMG_1396.jpg").resolve())
+    template_path = str(Path("./templates/CDC_card_template_01.png").resolve())
+    isValid = verify_card(img_path, template_path, fileTag='')
+
+    if isValid:
+        print("Valid vaccination card detected!")
+    else:
+        print("Invalid vaccination card detected!")
