+++ setup.cfg
@@ -1,3 +1,12 @@
-[flake8]
-max-line-length = 180
-import-order-style = smarkets
+[metadata]
+name = vcv
+version = 0.0.1
+
+[options]
+packages = find:
+install_requires =
+    imutils==0.5.4
+    numpy==1.20.1
+    opencv-python==4.5.1.48
+    Pillow==8.1.2
+    pytesseract==0.3.7
