+++ vcv/__init__.py
@@ -1,0 +1,5 @@
+"""A tool for vaccination card verification"""
+
+__version__ = "0.0.1"
+
+from .verify_card import verify_card
