+++ vcv/pyimagesearch/alignment/align_images.py
